refactor: split data.py god-module and decompose B(10) blocks (#871)
data.py was the largest module in src/ (808 LOC) with the worst-ranking
complexity blocks. This splits it and removes those hotspots:

- Move the frame-construction/interpolation free functions into a new
  src/jquantstats/_utils/_construction.py (_to_polars, _apply_null_strategy,
  interpolate, _subtract_risk_free, _require_date_col, _align_returns_benchmark,
  plus a new _prices_to_returns). data.py re-imports them, so
  `jquantstats.data._subtract_risk_free` and `from jquantstats import interpolate`
  keep working.
- Move the "derive a new Data" methods (resample, copy, head, tail, truncate and
  the _truncate_* helpers) into src/jquantstats/_data_reshape.py as _ReshapeMixin,
  which Data now inherits. The mixin carries empty __slots__ so the frozen
  slotted dataclass is unchanged; it rebuilds via a lazily-imported Data to
  avoid the accessor import cycle.
- Decompose the B(10)/B(8) blocks: _apply_null_strategy -> B(5) (extracted
  _value_columns/_columns_with_nulls), Data.from_prices -> A(2) (extracted
  _prices_to_returns), _truncate_integer -> A (extracted _resolve_row_bound).

Result: data.py drops 808 -> 453 LOC; no block in data.py exceeds B(6).
make test (100% coverage), typecheck (ty + mypy strict) and docs-coverage
(100%) all stay green.

Closes #871
